requirements.txt
Flask
gTTS
Pillow
pytesseract
SQLAlchemy
psycopg2-binary